Tightening Up Intro (1)
Cleaning up intro. Also adds references to Kyber in CMS and AUTHKEM as well as ASN.1 references.

I almost included that "Kyber is an IND-CCA2-secure key-encapsulation mechanism (KEM)", but figured we could into IND-CCA2 in the security considerations as few are going to have a clue about the significance of IND-CCA2.

NOTE: Also see #26, which is an alternative intro.